Implement function to calculate nearest patch.
This will fix issues on some maps, where the car is not starting on the road.
Also if you switch to AI mode with F9, when the AI is outside of the road,
it will find its way back.
Also this is fixing another issue too. When a car gets down from the road,
it tried to go back to the patch, where it was leaving the road. This is
usually far behind the current position, which will cause, that the AI
instead of just getting back to the road it will try to go back and face
a wall, since it wants to go in the opposite direction.