Implement recover mode.
Recover mode will basically detect walls on the front
of the car, then go reverse if needed for 3 secs,
then recheck if there is still a wall on the front. If there is no wall,
break to zero speed and then switch to first gear back and leave recover mode.