docs: generating Doxygen locally and new Sourceforge URL
The documentation for getting game data was slightly out of date so this
commit updates the new process for getting a snapshot.

As the online version of Doxygen code docs does not seem to be online,
instructions have been added for generating the docs locally. Doxyfile
was also updated using Doxygen's automated update: `doxygen -u`.

Added the doxygen folder to *.gitignore*. I am not sure if this breaks
any pipelines used by the project.